backend/requirements.txt
google-genai==1.60.0
python-telegram-bot==22.5
jinja2==3.1.4
python-dotenv==1.0.1
weasyprint==62.3

# FastAPI backend stack
fastapi==0.128.0
uvicorn[standard]==0.32.1
pydantic==2.10.2
pydantic-settings==2.6.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
asyncpg==0.31.0
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
pydantic==2.10.2
